Move IMessage interface to publisher for easier review
