+++ src/Foundatio/Messaging/IMessageSubscriber.cs
@@ -46,29 +46,6 @@ namespace Foundatio.Messaging {
         DateTime CreatedUtc { get; }
     }
 
-    public interface IMessage<T> : IMessage where T: class {
-        T Body { get; }
-    }
-
-    public interface IMessage {
-        // correlation id used in logging
-        string CorrelationId { get; }
-        // used for rpc (request/reply)
-        string ReplyTo { get; }
-        // message priority
-        int Priority { get; }
-        // topic the message will be sent to
-        string Topic { get; }
-        // message type, will be converted to string and stored with the message for deserialization
-        Type MessageType { get; }
-        // message body
-        object GetBody();
-        // when the message should expire
-        DateTime ExpiresAtUtc { get; }
-        // additional data to store with the message
-        DataDictionary Data { get; }
-    }
-
     public interface IMessageContext : IMessage, IMessagePublisher, IDisposable {
         // message id
         string Id { get; }
